+++ src/Lens.ts
@@ -190,7 +190,7 @@ export const component: <A extends ReadonlyArray<unknown>, P extends keyof A>(
  * @since 2.3.0
  */
 export const index = (i: number) => <S, A>(sa: Lens<S, ReadonlyArray<A>>): Optional<S, A> =>
-  pipe(sa, asOptional, _.optionalComposeOptional(_.indexArray<A>().index(i)))
+  pipe(sa, asOptional, _.optionalComposeOptional(_.indexReadonlyArray<A>().index(i)))
 
 /**
  * Return a `Optional` from a `Lens` focused on a `ReadonlyRecord` and a key
@@ -199,7 +199,7 @@ export const index = (i: number) => <S, A>(sa: Lens<S, ReadonlyArray<A>>): Optio
  * @since 2.3.0
  */
 export const key = (key: string) => <S, A>(sa: Lens<S, Readonly<Record<string, A>>>): Optional<S, A> =>
-  pipe(sa, asOptional, _.optionalComposeOptional(_.indexRecord<A>().index(key)))
+  pipe(sa, asOptional, _.optionalComposeOptional(_.indexReadonlyRecord<A>().index(key)))
 
 /**
  * Return a `Lens` from a `Lens` focused on a `ReadonlyRecord` and a required key